Move metrics to top-level module
